More error conversion and win64 fixes
I decided to fix the win64 warnings and in the process, it made
sense to convert several other files to the new error style,
namely: netops.c, pkt.c, pqueue.c, revwalk.c, transports/local.c.

This also touches lots of other files since it fixes most of the
win64 warnings.  There are a number of strange bits in that code,
including win32 using int as a parameter type for most socket calls
that take a byte length value (so I needed some new wrappers).

I did not fix the 64-bit warnings on status.c since that is
undergoing such a large change on the status-rewrite branch and
I did not want to generate conflicts.  Also, I fixed warnings in
the xdiff code but I did not fix a bunch of 64 bit warnings in the
deps/regex code.

+++ src/signature.c
@@ -47,7 +47,7 @@ static int signature_error(const char *msg)
 static int process_trimming(const char *input, char **storage, const char *input_end, int fail_when_empty)
 {
 	const char *left, *right;
-	int trimmed_input_length;
+	size_t trimmed_input_length;
 
 	assert(storage);
